♻️ サーバー側で実行する関数を集約
