--- src/actions/action.d.ts
@@ -1,12 +1,0 @@
-export type ServerActionsResponse =
-  | {
-      success: true;
-      status: 200;
-      data: any;
-      message?: string;
-    }
-  | {
-      success: false;
-      status: number;
-      message: string;
-    };
