+++ src/features/recipes/index.ts
@@ -7,11 +7,7 @@ export * from "./components/recipe-list/recipe-list-item";
 export * from "./components/recipe-detail/recipe-edit-drop-down-menu";
 
 // lib
-export * from "./lib/action";
-//TODO:ファイル名をget-recipesからget-recipe-listに変えたい
-export * from "./lib/get-recipes";
-export * from "./lib/get-recipe";
-export * from "./lib/get-recipe-favorite";
+export * from "./lib/actions";
 
 //type
-export * from "./types/";
+export * from "./types";
