+++ src/features/recipes/lib/actions.ts
@@ -1,9 +1,18 @@
-import { Selectable, sql } from "kysely";
+"use server";
+
+import { revalidatePath } from "next/cache";
+
+import { Selectable, sql, Updateable } from "kysely";
+import { jsonArrayFrom } from "kysely/helpers/mysql";
 
 import { DATE_SPAN_RECENT, RECIPE_COUNT_FAVORITED_RECENTLY } from "@/config";
+import { ERROR_MESSAGE_UNKOWN_ERROR } from "@/config/error-message";
 import { db } from "@/lib/kysely";
-import { RecipeFavorite } from "@/types/db";
+import { Recipe, RecipeFavorite } from "@/types/db";
 
+import { RecipeListItemType, ServerActionsResponse } from "../types";
+
+// get-recipes.ts
 export async function getRecipesWithFavoriteCount({
   query,
   page = 1,
@@ -211,4 +220,187 @@ async function getRecipesFavoriteCount() {
       return a.recipeFavoriteCount > b.recipeFavoriteCount ? -1 : 1;
     })
     .slice(0, RECIPE_COUNT_FAVORITED_RECENTLY);
+}
+
+// recipe-favorite.ts
+export async function updateRecipeFavorite(
+  recipeId: string,
+  userId: string,
+  isFavorite: boolean,
+): Promise<ServerActionsResponse> {
+  try {
+    await db
+      .updateTable("RecipeFavorite")
+      .set({
+        deletedAt: new Date(),
+      })
+      .where("recipeId", "=", recipeId)
+      .where("userId", "=", userId)
+      .execute();
+    if (isFavorite) {
+      console.log("isFavorite true");
+      const recipeFavorite = await db
+        .selectFrom("RecipeFavorite")
+        .select(["id"])
+        .where("recipeId", "=", recipeId)
+        .where("userId", "=", userId)
+        .orderBy("updatedAt", "desc")
+        .executeTakeFirst();
+
+      if (recipeFavorite) {
+        await db
+          .updateTable("RecipeFavorite")
+          .set({
+            updatedAt: new Date(),
+            deletedAt: null,
+          })
+          .where("id", "=", recipeFavorite.id)
+          .execute();
+      } else {
+        db.insertInto("RecipeFavorite")
+          .values({
+            recipeId: recipeId,
+            userId: userId,
+          })
+          .executeTakeFirst();
+      }
+    }
+
+    return {
+      success: true,
+      status: 200,
+      data: { recipeFavorite: { isFavorite } },
+    };
+  } catch (e) {
+    return {
+      success: false,
+      status: 500,
+      message: ERROR_MESSAGE_UNKOWN_ERROR,
+    };
+  } finally {
+    revalidatePath(`/recipe/${recipeId}`);
+  }
+}
+
+// get-recipe.ts
+export async function getRecipeById(id: string) {
+  const baseQuery = db
+    .selectFrom("Recipe")
+    .selectAll("Recipe")
+    .select((eb) => [
+      jsonArrayFrom(
+        eb
+          .selectFrom("User")
+          .select(["User.id", "User.name", "User.image"])
+          .whereRef("User.id", "=", "Recipe.userId")
+          .where("User.deletedAt", "is", null),
+      ).as("User"),
+      jsonArrayFrom(
+        eb
+          .selectFrom("RecipeImage")
+          .select(["RecipeImage.id", "RecipeImage.imgSrc", "RecipeImage.index"])
+          .whereRef("RecipeImage.recipeId", "=", "Recipe.id")
+          .where("RecipeImage.deletedAt", "is", null)
+          .orderBy("RecipeImage.index", "asc"),
+      ).as("RecipeImage"),
+      jsonArrayFrom(
+        eb
+          .selectFrom("RecipeIngredient")
+          .select(["RecipeIngredient.id", "RecipeIngredient.name", "RecipeIngredient.index"])
+          .whereRef("RecipeIngredient.recipeId", "=", "Recipe.id")
+          .where("RecipeIngredient.deletedAt", "is", null)
+          .orderBy("RecipeIngredient.index", "asc"),
+      ).as("RecipeIngredient"),
+      jsonArrayFrom(
+        eb
+          .selectFrom("RecipeCookingProcedure")
+          .select(["RecipeCookingProcedure.id", "RecipeCookingProcedure.name", "RecipeCookingProcedure.index"])
+          .whereRef("RecipeCookingProcedure.recipeId", "=", "Recipe.id")
+          .where("RecipeCookingProcedure.deletedAt", "is", null)
+          .orderBy("RecipeCookingProcedure.index", "asc"),
+      ).as("RecipeCookingProcedure"),
+      jsonArrayFrom(
+        eb
+          .selectFrom("RecipeLink")
+          .select(["RecipeLink.id", "RecipeLink.url", "RecipeLink.category", "RecipeLink.index"])
+          .whereRef("RecipeLink.recipeId", "=", "Recipe.id")
+          .where("RecipeLink.deletedAt", "is", null)
+          .orderBy("RecipeLink.index", "asc"),
+      ).as("RecipeLink"),
+    ])
+    .where("Recipe.deletedAt", "is", null);
+
+  return await baseQuery.where("Recipe.id", "=", id).executeTakeFirst();
+}
+
+// get-recipe-favorite.ts
+export async function getFavoriteCountByRecipeId(recipeId: string) {
+  const recipeFavorites: Pick<Selectable<RecipeFavorite>, "recipeId">[] = await db
+    .selectFrom("RecipeFavorite")
+    .select(["recipeId"])
+    .where("recipeId", "=", recipeId)
+    .where("deletedAt", "is", null)
+    .execute();
+
+  return recipeFavorites.length;
+}
+
+export async function getIsFavoriteByUserId(recipeId: string, userId: string) {
+  const recipeFavorites: Pick<Selectable<RecipeFavorite>, "recipeId">[] = await db
+    .selectFrom("RecipeFavorite")
+    .select(["recipeId", "userId"])
+    .where("recipeId", "=", recipeId)
+    .where("userId", "=", userId)
+    .where("deletedAt", "is", null)
+    .execute();
+
+  return recipeFavorites.length > 0;
+}
+
+// action.ts
+export async function fetchRecipesWithFavoriteCount({
+  search,
+  page = 1,
+}: {
+  search?: string | undefined;
+  page?: number;
+}) {
+  const recipes: RecipeListItemType = await getRecipesWithFavoriteCount({ query: search, page });
+  return recipes;
+}
+
+export async function fetchRecipesFavoritedRecently({
+  search,
+  page = 1,
+}: {
+  search?: string | undefined;
+  page?: number;
+}) {
+  const recipes: RecipeListItemType = await getRecipesFavoritedRecently({ query: search, page });
+  return recipes;
+}
+
+// recipe.ts
+export async function update(recipeId: string, updateValues: Updateable<Recipe>): Promise<ServerActionsResponse> {
+  try {
+    const result = await db.updateTable("Recipe").set(updateValues).where("id", "=", recipeId).executeTakeFirst();
+    if (result.numUpdatedRows !== BigInt("1")) {
+      throw new Error(ERROR_MESSAGE_UNKOWN_ERROR);
+    }
+    const updatedRecipe = await db.selectFrom("Recipe").selectAll().where("id", "=", recipeId).executeTakeFirst();
+
+    return {
+      success: true,
+      status: 200,
+      data: { recipe: updatedRecipe },
+    };
+  } catch (e) {
+    return {
+      success: false,
+      status: 500,
+      message: ERROR_MESSAGE_UNKOWN_ERROR,
+    };
+  } finally {
+    revalidatePath(`/recipe/${recipeId}`);
+  }
 }
