--- src/features/recipes/lib/get-recipe.ts
@@ -1,57 +1,0 @@
-import { jsonArrayFrom } from "kysely/helpers/mysql";
-
-import { db } from "@/lib/kysely";
-
-export async function getRecipeById(id: string) {
-  const baseQuery = createBaseQuerySelect();
-
-  return await baseQuery.where("Recipe.id", "=", id).executeTakeFirst();
-}
-
-function createBaseQuerySelect() {
-  return db
-    .selectFrom("Recipe")
-    .selectAll("Recipe")
-    .select((eb) => [
-      jsonArrayFrom(
-        eb
-          .selectFrom("User")
-          .select(["User.id", "User.name", "User.image"])
-          .whereRef("User.id", "=", "Recipe.userId")
-          .where("User.deletedAt", "is", null),
-      ).as("User"),
-      jsonArrayFrom(
-        eb
-          .selectFrom("RecipeImage")
-          .select(["RecipeImage.id", "RecipeImage.imgSrc", "RecipeImage.index"])
-          .whereRef("RecipeImage.recipeId", "=", "Recipe.id")
-          .where("RecipeImage.deletedAt", "is", null)
-          .orderBy("RecipeImage.index", "asc"),
-      ).as("RecipeImage"),
-      jsonArrayFrom(
-        eb
-          .selectFrom("RecipeIngredient")
-          .select(["RecipeIngredient.id", "RecipeIngredient.name", "RecipeIngredient.index"])
-          .whereRef("RecipeIngredient.recipeId", "=", "Recipe.id")
-          .where("RecipeIngredient.deletedAt", "is", null)
-          .orderBy("RecipeIngredient.index", "asc"),
-      ).as("RecipeIngredient"),
-      jsonArrayFrom(
-        eb
-          .selectFrom("RecipeCookingProcedure")
-          .select(["RecipeCookingProcedure.id", "RecipeCookingProcedure.name", "RecipeCookingProcedure.index"])
-          .whereRef("RecipeCookingProcedure.recipeId", "=", "Recipe.id")
-          .where("RecipeCookingProcedure.deletedAt", "is", null)
-          .orderBy("RecipeCookingProcedure.index", "asc"),
-      ).as("RecipeCookingProcedure"),
-      jsonArrayFrom(
-        eb
-          .selectFrom("RecipeLink")
-          .select(["RecipeLink.id", "RecipeLink.url", "RecipeLink.category", "RecipeLink.index"])
-          .whereRef("RecipeLink.recipeId", "=", "Recipe.id")
-          .where("RecipeLink.deletedAt", "is", null)
-          .orderBy("RecipeLink.index", "asc"),
-      ).as("RecipeLink"),
-    ])
-    .where("Recipe.deletedAt", "is", null);
-}
